# idt/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: IDT API
  version: "1.0.0"
  description: |
    FastAPI + LangGraph/LangChain 기반 RAG & Agent 시스템 API

    **아키텍처**: Thin DDD (Domain → Application → Infrastructure)

    **주요 기능**:
    - PDF/Excel 문서 업로드 및 벡터 색인
    - BM25 + Vector 하이브리드 검색
    - ReAct RAG Agent 기반 질의응답
    - Multi-Turn 대화 메모리 관리
    - Agent Builder (Custom / Middleware / Auto)
    - MCP 서버 동적 등록 관리
servers:
  - url: http://localhost:8000
    description: Local Development
...
paths:
...
  /api/v1/agents:
    post:
      tags: [Agent Builder]
      summary: 에이전트 생성
      description: LLM이 도구 자동 선택 + 시스템 프롬프트 자동 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRequest"
      responses:
        "201":
          description: 에이전트 생성 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAgentResponse"
  /api/v1/agents/{agent_id}:
    get:
      tags: [Agent Builder]
      summary: 에이전트 조회
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 에이전트 정의
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Agent Builder]
      summary: 에이전트 시스템 프롬프트/이름 수정
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentRequest"
      responses:
        "200":
          description: 수정 성공
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/agents/{agent_id}/run:
    post:
      tags: [Agent Builder]
      summary: 에이전트 실행
      description: DB에서 워크플로우 로드 → LangGraph 동적 컴파일 → 응답
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunAgentRequest"
      responses:
        "200":
          description: 실행 결과
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v2/agents:
    post:
      tags: [middleware-agent]
      summary: 미들웨어 에이전트 생성
      description: |
        LangChain Middleware 기반 에이전트를 생성합니다.
        지원 미들웨어: SummarizationMiddleware, PIIMiddleware,
        ToolRetryMiddleware, ModelCallLimitMiddleware, ModelFallbackMiddleware
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMiddlewareAgentRequest"
      responses:
        "201":
          description: 에이전트 생성 성공
  /api/v2/agents/{agent_id}:
    get:
      tags: [middleware-agent]
      summary: 미들웨어 에이전트 조회
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
        - name: request_id
          in: query
          schema:
            type: string
            default: no-request-id
      responses:
        "200":
          description: 에이전트 정의
    patch:
      tags: [middleware-agent]
      summary: 미들웨어 에이전트 수정
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMiddlewareAgentRequest"
      responses:
        "200":
          description: 수정 성공
  /api/v2/agents/{agent_id}/run:
    post:
      tags: [middleware-agent]
      summary: 미들웨어 에이전트 실행
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunMiddlewareAgentRequest"
      responses:
        "200":
          description: 실행 결과

  # ── Auto Agent Builder (v3) ────────────────────────────────────────────
...
  /api/v1/mcp-registry:
    post:
      tags: [MCP Registry]
      summary: MCP 서버 등록
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterMCPServerRequest"
            example:
              name: "my-mcp-server"
              description: "Custom MCP tool server"
              transport: "stdio"
              command: "python"
              args: ["-m", "my_mcp_server"]
              user_id: "user-001"
      responses:
        "201":
          description: 등록 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MCPServerResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
    get:
      tags: [MCP Registry]
      summary: MCP 서버 목록 조회
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: string
          description: 필터 (없으면 전체 조회)
      responses:
        "200":
          description: MCP 서버 목록
  /api/v1/mcp-registry/{id}:
    get:
      tags: [MCP Registry]
      summary: 특정 MCP 서버 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: MCP 서버 정보
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [MCP Registry]
      summary: MCP 서버 정보 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMCPServerRequest"
      responses:
        "200":
          description: 수정 성공
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [MCP Registry]
      summary: MCP 서버 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: 삭제 성공
        "404":
          $ref: "#/components/responses/NotFound"

# ── Components ──────────────────────────────────────────────────────────────
components:
  responses:
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          example:
            detail: "Not found"
    ValidationError:
      description: 유효성 검사 실패
      content:
        application/json:
          example:
            detail: "strategy_type must be one of: ['full_token', 'parent_child', 'semantic']"
...
  schemas:
...
    CreateAgentRequest:
      type: object
      required: [name, goal]
      properties:
        name:
          type: string
        goal:
          type: string
        tool_ids:
          type: array
          items:
            type: string
    CreateAgentResponse:
      type: object
      properties:
        agent_id:
          type: string
        name:
          type: string
        system_prompt:
          type: string
        tool_ids:
          type: array
          items:
            type: string
    UpdateAgentRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        system_prompt:
          type: string
          nullable: true
    RunAgentRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
        session_id:
          type: string
          nullable: true
...
    CreateMiddlewareAgentRequest:
      type: object
      required: [name, goal]
      properties:
        name:
          type: string
        goal:
          type: string
        tool_ids:
          type: array
          items:
            type: string
        middlewares:
          type: array
          items:
            type: string
          description: "SummarizationMiddleware, PIIMiddleware, ToolRetryMiddleware, ModelCallLimitMiddleware, ModelFallbackMiddleware"
    UpdateMiddlewareAgentRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        system_prompt:
          type: string
          nullable: true
        middlewares:
          type: array
          items:
            type: string
    RunMiddlewareAgentRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
        session_id:
          type: string
          nullable: true
...
    RegisterMCPServerRequest:
      type: object
      required: [name, transport, command, user_id]
      properties:
        name:
          type: string
        description:
          type: string
        transport:
          type: string
          enum: [stdio, sse, websocket]
        command:
          type: string
        args:
          type: array
          items:
            type: string
        user_id:
          type: string
    UpdateMCPServerRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        command:
          type: string
          nullable: true
        args:
          type: array
          items:
            type: string
    MCPServerResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        transport:
          type: string
        command:
          type: string
        args:
          type: array
          items:
            type: string
        user_id:
          type: string
        created_at:
          type: string
          format: date-time